requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
asyncio==3.4.3
aiohttp==3.8.5